ADS1299: output the data in real units

+++ flight/PiOS/inc/pios_ads1299.h
@@ -42,7 +42,7 @@ struct pios_ads1299_cfg {
 };
 
 struct pios_ads1299_data {
-	uint32_t channels[8];
+	float channels[8];
 };
 
 /* Public Functions */
